requirements.txt
kafka-python
requests
mysql-connector-python
python-dotenv